修复因无法找到声音/片头/片尾或logo其中一个或几个而使程序无法正常执行的问题

+++ yunxin.php
@@ -42,19 +42,24 @@ function getNeed($classify,$para,$hf=0){
 		$reArr = scandir($GLOBALS[$para].'/common');
 		unset($reArr[0],$reArr[1]);
 		$count = count($reArr);
-		$k = rand(2,$count+1);
-		$path = $GLOBALS[$para]."/common/".$reArr[$k];
-		$filename = $reArr[$k];
+		if($count>0){
+			$k = rand(2,$count+1);
+			$path = $GLOBALS[$para]."/common/".$reArr[$k];
+			$filename = $reArr[$k];
+		}else{
+			$path = '';
+			$filename = '';
+		}
 	}
 	if($hf==1&&strpos($filename, '.mp4')){
 		$cmd = 'ffmpeg -i '.$path.' -q 0 '.str_replace('.mp4', '.mpg', $path);
 		shell_exec($cmd);
-		$res['path'] = str_replace('.mp4', '.mpg', $path);
-		$res['filename'] = str_replace('.mp4', '.mpg', $reArr[$k]);
+		$res['path'] = isset($filename)?str_replace('.mp4', '.mpg', $path):'';
+		$res['filename'] = str_replace('.mp4', '.mpg', $filename);
 		unlink($path);
 	}else{
-		$res['path'] = $path;
-		$res['filename'] = $reArr[$k];
+		$res['path'] = isset($filename)?$path:'';
+		$res['filename'] = $filename;
 	}
 	return $res;
 }
@@ -71,10 +76,10 @@ function getConfig(){
 	foreach ($cof as $k => $v) {
 		$v = explode(";", $v);
 		$config[$v[0]]['classify'] = $v[0];
-		$config[$v[0]]['cut_max_time'] = $v[1]?$v[1]:$GLOBALS['cut_max_time'];
-		$config[$v[0]]['header_time'] = $v[2]?$v[2]:$GLOBALS['header_time'];
-		$config[$v[0]]['footer_time'] = $v[3]?$v[3]:$GLOBALS['footer_time'];
-		$config[$v[0]]['cut_video_time'] = $v[4]?$v[4]:$GLOBALS['cut_video_time'];
+		$config[$v[0]]['cut_max_time'] = (isset($v[1])&&$v[1])||$v[1]==0?$v[1]:$GLOBALS['cut_max_time'];
+		$config[$v[0]]['header_time'] = (isset($v[2])&&$v[2])||$v[2]==0?$v[2]:$GLOBALS['header_time'];
+		$config[$v[0]]['footer_time'] = (isset($v[3])&&$v[3])||$v[3]==0?$v[3]:$GLOBALS['footer_time'];
+		$config[$v[0]]['cut_video_time'] = (isset($v[4])&&$v[4])||$v[4]==0?$v[4]:$GLOBALS['cut_video_time'];
 	}
 	return $config;
 }
@@ -139,42 +144,93 @@ function cutVideo($classify,$filename){
 	if(!is_dir('cache')){mkdir('cache');}
 	$vpath = $GLOBALS['video']."/".$classify."/".$filename;         //视频路径
 	$outpath = $GLOBALS['video_out']."/".$classify."/".$filename;   //输出路径 
+	//获取配置文件
 	$config = getConfig();
+	//获取视频信息
 	$videoInfo = getV($vpath);
 	$start_time = $config[$classify]['header_time'];
 	$end_time = $videoInfo['duration']-$config[$classify]['header_time']-$config[$classify]['footer_time']; 
+
+	//获取音乐
+	echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取音乐").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
 	$mp3Arr = getNeed($classify,$GLOBALS['mp3']);
 	$mp3 = $mp3Arr['path'];
+
+	//获取片头
+	echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取片头").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+	$headerArr = getNeed($classify,$GLOBALS['header'],1);
+	if($headerArr['path']){
+		$header = $headerArr['path'].'|';
+	}else{
+		$header = '';
+	}
+	
+	//获取片尾
+	echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取片尾").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+	$footerArr = getNeed($classify,$GLOBALS['footer'],1);
+	if($footerArr['path']){
+		$footer = '|'.$footerArr['path'];
+	}else{
+		$footer = '';
+	}
+
+	//获取logo
+	echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取logo").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+	$logoArr = getNeed($classify,$GLOBALS['logo']);
+	$logo = $logoArr['path'];
+
 	//判断是否需要分段
 	if($videoInfo['duration'] >= $config[$classify]['cut_max_time'] && $config[$classify]['cut_max_time']!=0){
 
 	}else{
-		//去除片头片尾加音乐
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("开始去除片头片尾").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$cmd = 'ffmpeg -i '.$vpath.' -stream_loop -1 -i '.$mp3.' -map 0:v -ss '.$start_time.' -t '.$end_time.' -c:v h264 -map 1:a -c:a copy cache/'.$filename;
-		shell_exec($cmd);
-		//转换成mpg
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("转换成mpg").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$cmd1 = 'ffmpeg -i '.'cache/'.$filename.' -q 0 cache/'.$filename.'.mpg'; 
-		shell_exec($cmd1);
-		//获取片头
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取片头").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$headerArr = getNeed($classify,$GLOBALS['header'],1);
-		$header = $headerArr['path'];
-		//获取片尾
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("开始获取片尾").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$footerArr = getNeed($classify,$GLOBALS['footer'],1);
-		$footer = $footerArr['path'];
-		//添加片头片尾（合成视频）
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("开始添加片头片尾（合成视频）").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$cmd2 = 'ffmpeg -i "concat:'.$header.'|cache/'.$filename.'.mpg|'.$footer.'" -q:a 10 cache/out'.$filename;
-		shell_exec($cmd2);
-		//获取logo
-		echo "*********************".PHP_EOL.PHP_EOL.strTo("开始添加logo").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
-		$logoArr = getNeed($classify,$GLOBALS['logo']);
-		$logo = $logoArr['path'];
-		$cmd2 = 'ffmpeg -i "cache/out'.$filename.'" -i '.$logo.' -filter_complex "[1:v][0:v]scale2ref=150:150 [wm][base];[base][wm]overlay=main_w-overlay_w-10:main_h-overlay_h-10" '.$GLOBALS['video_out'].'/'.$classify.'/'.$filename;
-		shell_exec($cmd2);
+		//判断开始时间，结束时间
+		if((int)$start_time+(int)$end_time<$videoInfo['duration']){
+			$start_time = $start_time>0?' -ss '.$start_time:'';
+			$end_time = $config[$classify]['footer_time']>0?' -t '.$end_time:'';
+		}else{
+			$start_time = '';
+			$end_time = '';
+		}
+
+		//判断MP3是否存在
+		if($mp3){
+			//去除片头片尾加音乐
+			echo "*********************".PHP_EOL.PHP_EOL.strTo("开始去除片头片尾加背景音乐").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+			$cmd = 'ffmpeg -i '.$vpath.' -stream_loop -1 -i '.$mp3.' -map 0:v '.$start_time.$end_time.' -c:v h264 -map 1:a -c:a copy cache/'.$filename;
+			shell_exec($cmd);
+		}else{
+			if($start_time&&$end_time){
+				$cmd = 'ffmpeg '.$start_time.$end_time.' -i '.$vpath.' cache/'.$filename;
+				shell_exec($cmd);
+			}else{
+				rename($vpath, 'cache/'.$filename);
+			}
+		}
+
+		//判断是否有片头片尾
+		if(!$footer&&!$header){
+			echo strTo('没有可用的片头片尾').PHP_EOL;
+			$cmd2 = rename('cache/'.$filename, 'cache/out'.$filename);
+		}else{
+			//转换成mpg
+			echo "*********************".PHP_EOL.PHP_EOL.strTo("开始转换成mpg").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+			$cmd1 = 'ffmpeg -i '.'cache/'.$filename.' -q 0 cache/'.$filename.'.mpg'; 
+			shell_exec($cmd1);
+
+			//添加片头片尾（合成视频）
+			echo "*********************".PHP_EOL.PHP_EOL.strTo("开始添加片头片尾（合成视频）").PHP_EOL.PHP_EOL."*********************".PHP_EOL;
+			$cmd2 = 'ffmpeg -i "concat:'.$header.'cache/'.$filename.'.mpg'.$footer.'" -q:a 10 cache/out'.$filename;
+			shell_exec($cmd2);
+		}
+
+		//判断是否有logo
+		if($logo){
+			$cmd3 = 'ffmpeg -i "cache/out'.$filename.'" -i '.$logo.' -filter_complex "[1:v][0:v]scale2ref=150:150 [wm][base];[base][wm]overlay=main_w-overlay_w-10:main_h-overlay_h-10" '.$GLOBALS['video_out'].'/'.$classify.'/'.$filename;
+			shell_exec($cmd3);
+		}else{
+			$cmd3 = rename('cache/out'.$filename, $GLOBALS['video_out'].'/'.$classify.'/'.$filename);
+		}
+
 	}
 	$cache = scandir('cache');
 	foreach ($cache as $k => $v) {
@@ -182,10 +238,12 @@ function cutVideo($classify,$filename){
 			unlink('cache/'.$v);
 		}
 	}
-	if(is_file($GLOBALS['video_out'].'/'.$classify.'/'.$filename)){
+	if(is_file($GLOBALS['video_out'].'/'.$classify.'/'.$filename)&&is_file($GLOBALS['video'].'/'.$classify.'/'.$filename)){
 		unlink($vpath);
 	}
 }
+
+
 $path = scandir($GLOBALS['video']);
 foreach ($path as $k => $v) {
 	if($v!='.'&&$v!='..'){
